+++ R/r2_bayes.R
@@ -4,7 +4,8 @@
 #' @description Compute R2 for Bayesian models. For mixed models (including a random part),
 #' it additionally computes the R2 related to the fixed effects only (marginal R2).
 #'
-#' @param model A Bayesian regression model.
+#' @param model A Bayesian regression model (from \code{brms}, \code{rstanarm},
+#'   \code{BayesFactor}, etc).
 #' @param robust Logical, if \code{TRUE}, the median instead of mean is used to
 #'   calculate the central tendency of the variances.
 #' @param ci Value or vector of probability of the CI (between 0 and 1) to be estimated.
@@ -36,6 +37,24 @@
 #'   )
 #'   r2_bayes(model)
 #' }
+#'
+#'
+#' if (require(BayesFactor)) {
+#'   data(mtcars)
+#'
+#'   BFM <- generalTestBF(mpg ~ cyl * gear, data = mtcars, progress = FALSE)
+#'   FM <- lm(mpg ~ cyl * gear, data = mtcars)
+#'
+#'   r2(FM)
+#'   r2(BFM[4])
+#'
+#'
+#'   # with random effects:
+#'   mtcars$gear <- factor(mtcars$gear)
+#'   model <- lmBF(mpg ~ hp + cyl + gear + gear:wt, mtcars, progress = FALSE, whichRandom = c("gear", "gear:wt"))
+#'   r2(model)
+#' }
+#'
 #' \dontrun{
 #' if (require("brms")) {
 #'   model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
@@ -47,11 +66,16 @@
 #' }
 #' @references Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2018). R-squared for Bayesian regression models. The American Statistician, 1–6. \doi{10.1080/00031305.2018.1549100}
 #'
+#' @export
+r2_bayes <- function(model, robust = TRUE, ci = .89) {
+  UseMethod("r2_bayes")
+}
+
+#' @export
 #' @importFrom insight find_algorithm is_multivariate find_response
 #' @importFrom stats median mad sd
 #' @importFrom bayestestR ci
-#' @export
-r2_bayes <- function(model, robust = TRUE, ci = .89) {
+r2_bayes.default <- function(model, robust = TRUE, ci = .89) {
   r2_bayesian <- .r2_posterior(model)
 
   if (is.null(r2_bayesian)) {
@@ -78,6 +102,60 @@ r2_bayes <- function(model, robust = TRUE, ci = .89) {
     )
   }
 }
+
+
+#' @export
+#' @importFrom insight get_parameters get_response find_predictors
+#' @importFrom stats median mad sd
+#' @importFrom bayestestR point_estimate hdi
+r2_bayes.BFBayesFactor <- function(model, robust = TRUE, ci = .89, ...){
+  if (!requireNamespace("rstantools", quietly = TRUE)) {
+    stop("Package `rstantools` needed for this function to work. Please install it.")
+  }
+
+  if (!requireNamespace("BayesFactor", quietly = TRUE)) {
+    stop("Package `BayesFactor` needed for this function to work. Please install it.")
+  }
+
+  # Estimates
+  params <- insight::get_parameters(model, unreduce = FALSE)
+  # remove sig and g cols
+  params <- params[, !grepl(pattern = "^sig2$|^g_|^g$", colnames(params))]
+
+  # Model Matrix
+  mm <- BayesFactor::model.matrix(model[1])
+  colnames(mm)[1] <- "mu"
+
+  # match?
+  all(colnames(params) == colnames(mm))
+
+  # Compute R2!
+  y <- insight::get_response(model)
+  yy <- as.matrix(params) %*% t(mm)
+  r2s <- rstantools::bayes_R2(yy, y = y)
+  r2_bayesian <- data.frame(R2_Bayes = r2s)
+
+  rand <- insight::find_predictors(model, effects = "random", flatten = TRUE)
+  if (!is.null(rand)) {
+    idx <- sapply(paste0("\\b", rand, "\\b"), grepl, x = colnames(params))
+    idx <- apply(idx, 1, any)
+    params[idx] <- 0
+
+    yy <- as.matrix(params) %*% t(mm)
+    r2s_marginal <- rstantools::bayes_R2(yy, y = y)
+    r2_bayesian$R2_Bayes_marginal <- r2s_marginal
+  }
+
+  structure(
+    class = "r2_bayes",
+    lapply(r2_bayesian, ifelse(robust, stats::median, mean)),
+    "SE" = lapply(r2_bayesian, ifelse(robust, stats::mad, stats::sd)),
+    "Estimates" = lapply(r2_bayesian, bayestestR::point_estimate, centrality = "all", dispersion = TRUE),
+    "CI" = lapply(r2_bayesian, bayestestR::hdi, ci = ci),
+    "robust" = robust
+  )
+}
+